refactor cleanup: - job: combine crawl_updater logic into base crawl_job - job: simplify loading initial scale for crawl job - increment new 'crawlAttemptCount' in crawlconfig when crawl is started, not necessarily finished - remove unused methods
